requirements.txt
# Requirements
pyyaml